Fixed `allCases` to be a constant, not a variable, on NetworkInterface.Flags & NetworkInterface.Change.ModificationNature.
I'm surprised Swift allowed the use of a variable in this context to begin with, since logically `allCases` has to be a constant since on e.g. an enum the cases have to be defined at compile time.  Weird.

This was found through the Swift compiler's "get ready for version 6" concurrency warnings.