Added Linux support for the core NetworkInterfaceInfo module.
NetworkInterfaceChangeMonitoring is not currently supported on non-Apple platforms because it relies on Apple's Network package.  But presumably there is some equivalent for Linux (and Windows) that could be used instead.  Alas I have no knowledge of what's available in this regard.  Also, it's currently a bit too much of a stretch for me to invest the required time to research & develop a solution for Linux (or Windows).