Only include the 'metrics' functionality if Darwin can be imported.
It turns out Linux - or at least the random Linux distro used by GitHub Codespaces - lacks the necessary sysctl for retrieving the metrics (in fact Glibc appears to lack _everything_ regarding sysctl, including the call itself!).

At a glance it appears it's possible to get most of the basic traffic counters by manually reading /proc/net/dev, if someone in the future is willing to dig into that.