Conformed IPv4View to Equatable & Hashable.
Trivial to do since the underlying storage is just a UInt32.

I didn't think to do this earlier since the view is nominally just a transient implementation detail for accessing IPv4-specific properties of NetworkAddress, but it's turned out in practice it's useful to work with IPv4Views a bit more robustly (such that I'm now also wondering if it should just be called IPv4Address and formally made a first-class data structure in the taxonomy…).